pyproject.toml
[tool.poetry]
name = "consorcio-autogestionado-back"
version = "0.1.0"
description = "Backend del sistema Consorcio Autogestionado"
authors = []
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
psycopg2-binary = "^2.9.9"
supabase = "^2.5.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
bcrypt = "^4.1.0"
pydantic-settings = "^2.3.0"
pydantic = { extras = ["email"], version = "^2.7.0" }
python-multipart = "^0.0.9"
httpx = "^0.27.0"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
